collections tests split
