+++ tests/ws-requiring-tests.sh
@@ -11,8 +11,9 @@ then
 fi
 num_checks=0
 num_fails=0
-echo_and_demand_succeeds python tests/ws-tests/readonly/test_phylesystem_api.py
-echo_and_demand_succeeds python tests/ws-tests/readonly/test_amendments_api.py
+echo_and_demand_succeeds python tests/ws_tests/read_only/test_phylesystem_api.py
+echo_and_demand_succeeds python tests/ws_tests/read_only/test_amendments_api.py
+echo_and_demand_succeeds python tests/ws_tests/read_only/test_collections.py
 stf=0
 if test ${num_fails} -gt 0
 then
