+++ tests/ws_tests/read_only/test_phylesystem_api.py
@@ -7,13 +7,8 @@ from peyotl.utility import get_logger
 import unittest
 import requests
 import os
+
 _LOG = get_logger(__name__)
-from peyotl.phylesystem.helper import get_repos
-try:
-    get_repos()
-    HAS_LOCAL_PHYLESYSTEM_REPOS = True
-except:
-    HAS_LOCAL_PHYLESYSTEM_REPOS = False
 
 
 @unittest.skipIf('RUN_WEB_SERVICE_TESTS' not in os.environ,
@@ -26,26 +21,32 @@ class TestPhylesystemAPI(unittest.TestCase):
     def testRemoteTransSugar(self):
         pa = PhylesystemAPI(self.domains, get_from='api', transform='server')
         test_phylesystem_api_for_study(self, pa)
+
     def testRemoteStudyList(self):
         pa = PhylesystemAPI(self.domains, get_from='api')
         sl = pa.study_list
         # dev/production repos should have hundreds of studies
         self.assertTrue(len(sl) > 100)
+
     def testPushFailureState(self):
         pa = PhylesystemAPI(self.domains, get_from='api')
         sl = pa.push_failure_state
         self.assertTrue(sl[0] is True)
+
     def testRemoteSugar(self):
         pa = PhylesystemAPI(self.domains, get_from='api')
         test_phylesystem_api_for_study(self, pa)
+
     def testExternalSugar(self):
         pa = PhylesystemAPI(self.domains, get_from='external')
         test_phylesystem_api_for_study(self, pa)
+
     def testConfig(self):
         pa = PhylesystemAPI(self.domains, get_from='api')
         x = pa.phylesystem_config
         self.assertTrue(('repo_nexml2json' in x.keys()) or ('assumed_doc_version' in x.keys()))
         # TODO: remove 'assumed_doc_version' once the preset API domain has newer code
+
     def testExternalURL(self):
         # N.B. that the URL for this API call is an odd one, e.g.
         #    http://devapi.opentreeoflife.org/phylesystem/external_url/pg_10
@@ -55,6 +56,6 @@ class TestPhylesystemAPI(unittest.TestCase):
         sid = find_val_literal_meta_first(re['nexml'], 'ot:studyId', detect_nexson_version(re))
         self.assertTrue(sid in ['10', 'pg_10'])
 
+
 if __name__ == "__main__":
     unittest.main()
-
