+++ tests/ws_tests/test_taxomachine.py
